Add files via upload
Given a function F(X), with F = { f_{1}(x, y, z), f_{2}(x, y, z), f_{3}(x, y, z) }^{T}, X = {x, y, z} Find the roots of the non-linear system of equations F(X) = 0. Methods include Newton-Raphson, Broyden with Sherman-Morrison Inverse, Gradient Search and combination Gradient + Broyden Search. The code also includes an implementation of LU decomposition dor solving a system of linear equations. The code is capable of handling systems in 2 and 3 dimensions as it is written, with a bit of work it could extended to systems of n >= 3 equations, all that's required is a new constructor.